GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Add squares contract (refs/gitbutler/Add-squares-contract)
   branch head: 386bbc0c8bd25522a6c95d4f4a559482023350fe
   - contracts/squares/square.ral
   - contracts/squares/squares.ral

Your previous branch was: refs/heads/swaps-and-auctions

The sha for that commit was: 0f2104ea2ada62e3a62185632806d68026ac8404

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch